Add attributes for C# required properties to MS.ASPCore.Razor.Common
