+++ src/Razor/src/Microsoft.AspNetCore.Razor.Common/LanguageSupport/CompilerFeatureRequiredAttribute.cs
@@ -1,0 +1,32 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT license. See License.txt in the project root for license information.
+
+// Copied from https://github/dotnet/runtime
+
+#if !NET7_0_OR_GREATER
+
+namespace System.Runtime.CompilerServices;
+
+/// <summary>
+/// Indicates that compiler support for a particular feature is required for the location where this attribute is applied.
+/// </summary>
+[AttributeUsage(AttributeTargets.All, AllowMultiple = true, Inherited = false)]
+internal sealed class CompilerFeatureRequiredAttribute : Attribute
+{
+    public CompilerFeatureRequiredAttribute(string featureName)
+    {
+        FeatureName = featureName;
+    }
+
+    /// <summary>
+    /// The name of the compiler feature.
+    /// </summary>
+    public string FeatureName { get; }
+
+    /// <summary>
+    /// If true, the compiler can choose to allow access to the location where this attribute is applied if it does not understand <see cref="FeatureName"/>.
+    /// </summary>
+    public bool IsOptional { get; init; }
+}
+
+#endif
