+++ src/Razor/src/Microsoft.AspNetCore.Razor.Common/LanguageSupport/RequiredMemberAttribute.cs
@@ -1,0 +1,14 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT license. See License.txt in the project root for license information.
+
+// Copied from https://github/dotnet/runtime
+
+#if !NET7_0_OR_GREATER
+
+namespace System.Runtime.CompilerServices;
+
+/// <summary>Specifies that a type has required members or that a member is required.</summary>
+[AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Field | AttributeTargets.Property, AllowMultiple = false, Inherited = false)]
+internal sealed class RequiredMemberAttribute : Attribute { }
+
+#endif
