+++ src/Razor/src/Microsoft.AspNetCore.Razor.Common/LanguageSupport/SetsRequiredMembersAttribute.cs
@@ -1,0 +1,17 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT license. See License.txt in the project root for license information.
+
+// Copied from https://github/dotnet/runtime
+
+#if !NET7_0_OR_GREATER
+
+namespace System.Diagnostics.CodeAnalysis;
+
+/// <summary>
+/// Specifies that this constructor sets all required members for the current type, and callers
+/// do not need to set any required members themselves.
+/// </summary>
+[AttributeUsage(AttributeTargets.Constructor, AllowMultiple = false, Inherited = false)]
+internal sealed class SetsRequiredMembersAttribute : Attribute { }
+
+#endif
