--- src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/RequiredAttributes.cs
@@ -1,37 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT license. See License.txt in the project root for license information.
-
-#if !NETCOREAPP
-namespace System.Runtime.CompilerServices
-{
-    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Field | AttributeTargets.Property, Inherited = false, AllowMultiple = false)]
-    internal sealed class RequiredMemberAttribute : Attribute
-    {
-        public RequiredMemberAttribute()
-        {
-        }
-    }
-
-    [AttributeUsage(AttributeTargets.All, AllowMultiple = true, Inherited = false)]
-    internal sealed class CompilerFeatureRequiredAttribute : Attribute
-    {
-        public CompilerFeatureRequiredAttribute(string featureName)
-        {
-            FeatureName = featureName;
-        }
-        public string FeatureName { get; }
-        public bool IsOptional { get; set; }
-    }
-}
-
-namespace System.Diagnostics.CodeAnalysis
-{
-    [AttributeUsage(AttributeTargets.Constructor, Inherited = false, AllowMultiple = false)]
-    internal sealed class SetsRequiredMembersAttribute : Attribute
-    {
-        public SetsRequiredMembersAttribute()
-        {
-        }
-    }
-}
-#endif
